# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cotonou Municipal Garage API",
        "description": "API documentation for the Cotonou Municipal Garage System",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/payments": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get list of payments",
                "description": "Returns list of payments",
                "operationId": "getPaymentsList",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by reference or license plate",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "payment_method",
                        "in": "query",
                        "description": "Filter by payment method",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Filter by start date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Filter by end date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PaymentResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Store new payment",
                "description": "Returns payment data",
                "operationId": "storePayment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePaymentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/payments/{id}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment information",
                "description": "Returns payment data",
                "operationId": "getPaymentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of payment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StorePaymentRequest": {
                "title": "Store Payment Request",
                "required": [
                    "procedure_id",
                    "amount",
                    "payment_method"
                ],
                "properties": {
                    "procedure_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 75.5
                    },
                    "payment_method": {
                        "type": "string",
                        "enum": [
                            "cash",
                            "card",
                            "mobile_money"
                        ],
                        "example": "cash"
                    },
                    "reference": {
                        "type": "string",
                        "example": "TXN_123456789",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PaymentResource": {
                "title": "Payment Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "procedure_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 75.5
                    },
                    "payment_method": {
                        "type": "string",
                        "enum": [
                            "cash",
                            "card",
                            "mobile_money"
                        ],
                        "example": "cash"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "completed",
                            "failed",
                            "refunded"
                        ],
                        "example": "completed"
                    },
                    "reference": {
                        "type": "string",
                        "example": "TXN_123456789"
                    },
                    "paid_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "procedure": {
                        "$ref": "#/components/schemas/ProcedureResource"
                    },
                    "processed_by": {
                        "$ref": "#/components/schemas/UserResource"
                    }
                },
                "type": "object"
            },
            "ProcedureDocumentResource": {
                "title": "Procedure Document Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "procedure_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "document_type": {
                        "type": "string",
                        "example": "impound_report"
                    },
                    "file_path": {
                        "type": "string",
                        "example": "documents/impound_report_1.pdf"
                    },
                    "url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://localhost/storage/documents/impound_report_1.pdf"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ProcedureResource": {
                "title": "Procedure Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "vehicle_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "procedure_type": {
                        "type": "string",
                        "enum": [
                            "impound",
                            "release",
                            "auction"
                        ],
                        "example": "impound"
                    },
                    "description": {
                        "type": "string",
                        "example": "Vehicle impounded for illegal parking."
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "in_progress",
                            "completed",
                            "cancelled"
                        ],
                        "example": "pending"
                    },
                    "cost": {
                        "type": "number",
                        "format": "float",
                        "example": 75.5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_by": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "documents": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProcedureDocumentResource"
                        }
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "User Resource",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "ulid",
                        "example": "01H8XGJWBWBAQ8PEJ9C3J745A1"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "manager",
                            "agent"
                        ],
                        "example": "agent"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}